HarvestEstimateHandler.kt
FieldHarvestEstimate() and PlantatinoHarvestEstimate() both can get plants with 0 harvest but it's impossible for fields by design.

FieldHarvestEstimate()
Uses endHarvest != initialHarvest to log change now.
setting bools to null and killing the plant if harvest dropped to 0 is moved out of the log clause

PlantatinoHarvestEstimate()
Uses endHarvest != initialHarvest to log change now.
setting bools to null and killing the plant if drought is moved out of the log clause because it's still possible for a plantation with 0 harvest to be hit with drought.